requirements.txt
edge_tts==6.1.12
fastapi==0.115.0
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
rvc_python==0.1.4
uvicorn==0.30.6
